doc/tut2.txt: clarified usage of static/dynamic binding
The original text was slightly misleading in that it could be read that newLit and newPlus could not be implemented as methods; 
this is untrue, because newPlus could also be impleented as a method. The reason that newLit can't be implemented as a
method is that its first argument (x: int) has a non-object type.