Corrected gcc path on windows
When gcc path is set to  r"$nimrod\dist\mingw\bin", the compiler gives
an error:
Error : unhandled exception : invalid format string [Value Error], but
works correctly with gcc.path set to r"$nim\dist\mingw\bin". I think
this issue was caused due to the name change from nimrod to nim , but
the name change was not replicated in the config file.